Update from master

+++ HomeGenie/Service/Logging/StatisticsLogger.cs
@@ -40,6 +40,7 @@ namespace HomeGenie.Service.Logging
         public DateTime TimeEnd;
         public string Domain;
         public string Address;
+        public string CustomData;
         public double Value;
     }
 
@@ -47,7 +48,6 @@ namespace HomeGenie.Service.Logging
     public class StatisticsLogger
     {
         public static List<string> StatisticsFields = new List<string>() {
-            "Conditions.",
             "Sensor.",
             "Meter.",
             "PowerMonitor.",
@@ -71,9 +71,9 @@ namespace HomeGenie.Service.Logging
         private Timer logInterval;
         private HomeGenieService homegenie;
         private SQLiteConnection dbConnection;
+        private long dbSizeLimit = 2097152;
 
         //private object dbLock = new object();
-        private readonly long dbSizeLimit = 5242880;
         //private static int STATISTICS_TIME_RESOLUTION_MINUTES = 5;
         private readonly int _statisticsTimeResolutionSeconds = 5 * 60;
 
@@ -83,27 +83,61 @@ namespace HomeGenie.Service.Logging
             dbSizeLimit = hg.SystemConfiguration.HomeGenie.Statistics.MaxDatabaseSizeMBytes * 1024 * 1024;
             _statisticsTimeResolutionSeconds = hg.SystemConfiguration.HomeGenie.Statistics.StatisticsTimeResolutionSeconds;
 
-            logInterval = new Timer(TimeSpan.FromSeconds(_statisticsTimeResolutionSeconds).TotalMilliseconds);
-            logInterval.Elapsed += logInterval_Elapsed;
         }
 
+        /// <summary>
+        /// Start this instance.
+        /// </summary>
         public void Start()
         {
-            logInterval.Start();
             OpenStatisticsDatabase();
+            if (logInterval == null)
+            {
+                logInterval = new Timer(TimeSpan.FromSeconds(_statisticsTimeResolutionSeconds).TotalMilliseconds);
+                logInterval.Elapsed += logInterval_Elapsed;
+            }
+            logInterval.Start();
         }
 
+        /// <summary>
+        /// Stop this instance.
+        /// </summary>
         public void Stop()
         {
-            logInterval.Stop();
+            if (logInterval != null)
+            {
+                logInterval.Elapsed -= logInterval_Elapsed;
+                logInterval.Stop();
+                logInterval.Dispose();
+                logInterval = null;
+            }
             CloseStatisticsDatabase();
         }
 
-        public void DatabaseReset()
+        /// <summary>
+        /// Gets or sets the size limit.
+        /// </summary>
+        /// <value>The size limit.</value>
+        public long SizeLimit
+        {
+            get { return dbSizeLimit; }
+            set { dbSizeLimit = value; }
+        }
+
+        /// <summary>
+        /// Resets the database.
+        /// </summary>
+        public void ResetDatabase()
         {
             ResetStatisticsDatabase();
         }
 
+        /// <summary>
+        /// Gets the parameters list.
+        /// </summary>
+        /// <returns>The parameters list.</returns>
+        /// <param name="domain">Domain.</param>
+        /// <param name="address">Address.</param>
         public List<string> GetParametersList(string domain, string address)
         {
             var parameterList = new List<string>();
@@ -128,6 +162,10 @@ namespace HomeGenie.Service.Logging
             return parameterList;
         }
 
+        /// <summary>
+        /// Gets the date range.
+        /// </summary>
+        /// <returns>The date range.</returns>
         public StatisticsEntry GetDateRange()
         {
             var start = DateTime.UtcNow;
@@ -156,6 +194,12 @@ namespace HomeGenie.Service.Logging
             return new StatisticsEntry() { TimeStart = start, TimeEnd = end };
         }
 
+        /// <summary>
+        /// Gets the total counter.
+        /// </summary>
+        /// <returns>The total counter.</returns>
+        /// <param name="parameterName">Parameter name.</param>
+        /// <param name="timeScaleSeconds">Time scale seconds.</param>
         public double GetTotalCounter(string parameterName, double timeScaleSeconds)
         {
             double value = 0;
@@ -185,6 +229,16 @@ namespace HomeGenie.Service.Logging
             return value;
         }
 
+        /// <summary>
+        /// Gets the hourly counter.
+        /// </summary>
+        /// <returns>The hourly counter.</returns>
+        /// <param name="domain">Domain.</param>
+        /// <param name="address">Address.</param>
+        /// <param name="parameterName">Parameter name.</param>
+        /// <param name="timescaleseconds">Timescaleseconds.</param>
+        /// <param name="startDate">Start date.</param>
+        /// <param name="endDate">End date.</param>
         public List<StatisticsEntry> GetHourlyCounter(
             string domain,
             string address,
@@ -302,6 +356,14 @@ namespace HomeGenie.Service.Logging
             return values;
         }
 
+        /// <summary>
+        /// Gets the today detail.
+        /// </summary>
+        /// <returns>The today detail.</returns>
+        /// <param name="domain">Domain.</param>
+        /// <param name="address">Address.</param>
+        /// <param name="parameterName">Parameter name.</param>
+        /// <param name="aggregator">Aggregator.</param>
         public List<StatisticsEntry> GetTodayDetail(
             string domain,
             string address,
@@ -407,7 +469,16 @@ namespace HomeGenie.Service.Logging
                     dbCommand.Parameters.Add(new SQLiteParameter("@domain", domain));
                     dbCommand.Parameters.Add(new SQLiteParameter("@address", address));
                 }
-                string query = "select TimeStart,TimeEnd,Domain,Address," + aggregator + "(AverageValue) as Value from ValuesHist where " + filter + " Parameter = @parameterName AND " + GetParameterizedDateRangeFilter(ref dbCommand, startDate, endDate) + " group by Domain, Address, strftime('%H', TimeStart)  order by TimeStart asc;";
+                string query = "";
+                if (aggregator != "")
+                {
+                    if(aggregator == "All")
+                        query = "select TimeStart,TimeEnd,Domain,Address,CustomData,AverageValue from ValuesHist where " + filter + " Parameter = @parameterName AND " + GetParameterizedDateRangeFilter(ref dbCommand, startDate, endDate) + " order by CustomData, Domain, Address, TimeStart asc;";
+                    else
+                        query = "select TimeStart,TimeEnd,Domain,Address,CustomData," + aggregator + "(AverageValue) as Value from ValuesHist where " + filter + " Parameter = @parameterName AND " + GetParameterizedDateRangeFilter(ref dbCommand, startDate, endDate) + " group by Domain, Address, strftime('%H', TimeStart)  order by TimeStart asc;";
+                }
+                else
+                    query = "select TimeStart,TimeEnd,Domain,Address,CustomData,AverageValue from ValuesHist where " + filter + " Parameter = @parameterName AND " + GetParameterizedDateRangeFilter(ref dbCommand, startDate, endDate) + " order by TimeStart asc;";
                 dbCommand.Parameters.Add(new SQLiteParameter("@parameterName", parameterName));
 
                 //if (domain != "" && address != "") filter = "Domain ='" + domain + "' and Address = '" + address + "' and ";
@@ -422,18 +493,19 @@ namespace HomeGenie.Service.Logging
                     entry.TimeEnd = DateTime.Parse(reader.GetString(1));
                     entry.Domain = reader.GetString(2);
                     entry.Address = reader.GetString(3);
+                    entry.CustomData = reader.GetString(4);
                     entry.Value = 0;
                     try
                     {
-                        entry.Value = (double)reader.GetFloat(4);
+                        entry.Value = (double)reader.GetFloat(5);
                     }
                     catch
                     {
-                        var value = reader.GetValue(4);
+                        var value = reader.GetValue(5);
                         if (value != DBNull.Value && value != null) double.TryParse(
-                                reader.GetString(4),
-                                out entry.Value
-                            );
+                            reader.GetString(5),
+                            out entry.Value
+                        );
                     }
                     //
                     values.Add(entry);
@@ -444,10 +516,37 @@ namespace HomeGenie.Service.Logging
             return values;
         }
 
+        /// <summary>
+        /// Deletes the stat.
+        /// </summary>
+        /// <returns>The stat.</returns>
+        /// <param name="startDate">Start date.</param>
+        /// <param name="value">Value.</param>
+        public string DeleteStat(
+          DateTime startDate,
+            string value
+        )
+        {
+            var valueText = "";
+            double fValue = double.Parse(value,System.Globalization.CultureInfo.InvariantCulture);
+            var valueSplit = fValue.ToString("0.00", System.Globalization.CultureInfo.InvariantCulture);
+            if (valueSplit == value)
+                valueText = " AND (AverageValue = '" + value + "')";
+            var dbCommand = dbConnection.CreateCommand();
+            dbCommand.CommandText = "DELETE FROM ValuesHist WHERE (TimeStart BETWEEN '"+startDate.ToString("yyyy-MM-dd HH:mm:ss")+"' AND '"+startDate.ToString("yyyy-MM-dd HH:mm:ss")+".999999')"+valueText;
+            dbCommand.ExecuteNonQuery();
+            return dbCommand.CommandText;
+        }
+
+        /// <summary>
+        /// Opens the statistics database.
+        /// </summary>
+        /// <returns><c>true</c>, if statistics database was opened, <c>false</c> otherwise.</returns>
         private bool OpenStatisticsDatabase()
         {
             bool success = false;
             //lock (dbLock)
+            if (dbConnection == null)
             {
                 try
                 {
@@ -455,17 +554,27 @@ namespace HomeGenie.Service.Logging
                     dbConnection.Open();
                     success = true;
                 }
-                catch (Exception)
+                catch (Exception ex)
                 {
-                    // TODO: add error logging
+                    HomeGenieService.LogError(
+                        Domains.HomeAutomation_HomeGenie, 
+                        "Service.StatisticsLogger", 
+                        "Database Error", 
+                        "Exception.StackTrace", 
+                        String.Format("{0}: {1}", ex.Message, ex.StackTrace)
+                    );
                 }
             }
             return success;
         }
 
+        /// <summary>
+        /// Resets the statistics database.
+        /// </summary>
         private void ResetStatisticsDatabase()
         {
             //lock (dbLock)
+            if (dbConnection != null)
             {
                 var dbCommand = dbConnection.CreateCommand();
                 dbCommand.CommandText = "DELETE FROM ValuesHist";
@@ -474,6 +583,7 @@ namespace HomeGenie.Service.Logging
                 dbCommand.ExecuteNonQuery();
             }
         }
+
         /// <summary>
         /// Removes older values to keep DB size within configured size limit. Currently just cuts out last half of dates.
         /// </summary>
@@ -508,7 +618,7 @@ namespace HomeGenie.Service.Logging
                     dbCommand.CommandText = "VACUUM";
                     dbCommand.ExecuteNonQuery();
 
-                    HomeGenieService.LogEvent(
+                    HomeGenieService.LogDebug(
                                        Domains.HomeAutomation_HomeGenie,
                                        "Service.StatisticsLogger",
                                        "Cleaned old values from database.",
@@ -518,31 +628,55 @@ namespace HomeGenie.Service.Logging
                 }
             }
         }
+
+        /// <summary>
+        /// Closes the statistics database.
+        /// </summary>
         private void CloseStatisticsDatabase()
         {
             //lock (dbLock)
+            if (dbConnection != null)
             {
                 dbConnection.Close();
+                dbConnection.Dispose();
+                dbConnection = null;
             }
         }
 
+        /// <summary>
+        /// Gets the date range filter.
+        /// </summary>
+        /// <returns>The date range filter.</returns>
+        /// <param name="start">Start.</param>
+        /// <param name="end">End.</param>
         private string GetDateRangeFilter(DateTime start, DateTime end)
         {
-            var d1 = DateTime.Parse(start.ToString("yyyy-MM-dd") + " 00:00:00.000000");
-            var d2 = DateTime.Parse(end.ToString("yyyy-MM-dd") + " 23:59:59.999999");
+            var d1 = DateTime.Parse(start.ToLocalTime().ToString("yyyy-MM-dd") + " 00:00:00.000000");
+            var d2 = DateTime.Parse(end.ToLocalTime().ToString("yyyy-MM-dd") + " 23:59:59.999999");
             var filter = "(TimeStart >= '" + d1.ToUniversalTime().ToString("yyyy-MM-dd HH:mm:ss.ffffff") + "' AND TimeStart <= '" + d2.ToUniversalTime().ToString("yyyy-MM-dd HH:mm:ss.ffffff") + "')";
             return filter;
         }
 
+        /// <summary>
+        /// Gets the parameterized date range filter.
+        /// </summary>
+        /// <returns>The parameterized date range filter.</returns>
+        /// <param name="dbCommand">Db command.</param>
+        /// <param name="start">Start.</param>
+        /// <param name="end">End.</param>
         private string GetParameterizedDateRangeFilter(ref SQLiteCommand dbCommand, DateTime start, DateTime end)
         {
-            var d1 = DateTime.Parse(start.ToString("yyyy-MM-dd") + " 00:00:00.000000");
-            var d2 = DateTime.Parse(end.ToString("yyyy-MM-dd") + " 23:59:59.999999");
+            var d1 = DateTime.Parse(start.ToLocalTime().ToString("yyyy-MM-dd") + " 00:00:00.000000");
+            var d2 = DateTime.Parse(end.ToLocalTime().ToString("yyyy-MM-dd") + " 23:59:59.999999");
             dbCommand.Parameters.Add(new SQLiteParameter("@timeStartMin", d1.ToUniversalTime().ToString("yyyy-MM-dd HH:mm:ss.ffffff")));
             dbCommand.Parameters.Add(new SQLiteParameter("@timeStartMax", d2.ToUniversalTime().ToString("yyyy-MM-dd HH:mm:ss.ffffff")));
             return "(TimeStart >= @timeStartMin  AND TimeStart <= @timeStartMax)";
         }
 
+        /// <summary>
+        /// Gets the name of the statistics database.
+        /// </summary>
+        /// <returns>The statistics database name.</returns>
         private string GetStatisticsDatabaseName()
         {
             return Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "homegenie_stats.db");
@@ -553,15 +687,16 @@ namespace HomeGenie.Service.Logging
             return datetime.ToString("yyyy-MM-dd HH:mm:ss.ffffff");
         }
 
-
         private void logInterval_Elapsed(object sender, ElapsedEventArgs eventArgs)
         {
             var end = DateTime.UtcNow;
-            var modules = (TsList<Module>)homegenie.Modules; //.Clone();
-            foreach (var module in modules)
+            var modules = (TsList<Module>)homegenie.Modules; //.DeepClone();
+            for (int m = 0; m < modules.Count; m++)
             {
-                foreach (var parameter in module.Properties)
+                var module = modules[m];
+                for (int p = 0; p < module.Properties.Count; p++)
                 {
+                    var parameter = module.Properties[p];
                     if (parameter.Statistics.Values.Count > 0)
                     {
                         var values = parameter.Statistics.Values.FindAll(sv => (sv.Timestamp.Ticks <= end.Ticks && sv.Timestamp.Ticks > parameter.Statistics.LastProcessedTimestap.Ticks));
@@ -569,19 +704,19 @@ namespace HomeGenie.Service.Logging
                         {
                             double average = (values.Sum(d => d.Value) / values.Count);
                             //
-                            //TODO: check db file age/size for archiving old data
-                            //
-                            string dbName = GetStatisticsDatabaseName();
-                            var fileInfo = new FileInfo(dbName);
-                            if (fileInfo.Length > dbSizeLimit) // 5Mb limit for stats - temporary limitations to get rid of in the future
-                            {
-                                //ResetStatisticsDatabase();
-                                // TODO: Test method below, then use that instead of rsetting whole database.
-                                CleanOldValuesFromStatisticsDatabase();
-                            }
+                            //TODO: improve db file age/size check for archiving old data
                             //
                             try
                             {
+                                string dbName = GetStatisticsDatabaseName();
+                                var fileInfo = new FileInfo(dbName);
+                                if (fileInfo.Length > dbSizeLimit)
+                                {
+                                    ResetStatisticsDatabase();
+                                    // TODO: Test method below, then use that instead of rsetting whole database.
+                                    //CleanOldValuesFromStatisticsDatabase();
+                                }
+                                //
                                 var dbCommand = dbConnection.CreateCommand();
                                 // "TimeStart","TimeEnd","Domain","Address","Parameter","AverageValue", "CustomData"
                                 dbCommand.Parameters.Add(new SQLiteParameter("@timestart", DateTimeToSQLite(parameter.Statistics.LastProcessedTimestap)));
@@ -596,18 +731,22 @@ namespace HomeGenie.Service.Logging
                             }
                             catch (Exception ex)
                             {
-                                HomeGenieService.LogEvent(
-                                    Domains.HomeAutomation_HomeGenie,
-                                    "Service.StatisticsLogger",
-                                    ex.Message,
-                                    "Exception.StackTrace",
-                                    ex.StackTrace
+                                HomeGenieService.LogError(
+                                    Domains.HomeAutomation_HomeGenie, 
+                                    "Service.StatisticsLogger", 
+                                    "Database Error", 
+                                    "Exception.StackTrace", 
+                                    String.Format("{0}: {1}", ex.Message, ex.StackTrace)
                                 );
+                                // try close/reopen (perhaps some locking issue)
+                                CloseStatisticsDatabase();
+                                OpenStatisticsDatabase();
                             }
+                            //
+                            // reset statistics history sample
                             //
                             parameter.Statistics.LastProcessedTimestap = end;
                             parameter.Statistics.Values.Clear();
-
                         }
                     }
                 }
